refactor(back-office): split data-quality page into focused modules
pages/teams/data-quality/
  constants.ts    — FIELD_KEYS, FIELD_LABELS
  utils.ts        — formatFieldContent
  TeamLogoCell.tsx — logo + placeholder cell component
  ReviewModal.tsx  — enrichment review modal with field toggles

data-quality.tsx is now only state, hooks, table rendering,
and getServerSideProps
